generator/requirements.txt
gradio
langchain
langchain-google-genai
datasets
Pillow
matplotlib
numpy
